requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.0
django-environ==0.9.0
djangorestframework==3.13.1
environ==1.0
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
xlwt==1.3.0
